Removed unused ApplicationScreen.
Removed the `Application` entry from the `Screen` enum and removed the unused `ApplicationScreen` composable and related functions and imports in `HomeScreen` and the bottom sheet.

Added `appDrawerColumns` to `UserData` and the corresponding proto definition and serializer defaults. Also added `appDrawerColumns` as a parameter to `ApplicationScreen` and updated the `PagerScreen` to pass this value down.