Renamed `movingGridItem` to `moving` in `resolveConflictsWhenMoving` and `resolveConflicts`
The parameter `movingGridItem` in the `resolveConflictsWhenMoving` and `resolveConflicts` functions within the `MoveAlgorithm.kt` file has been renamed to `moving` for brevity and clarity. This change is also reflected in the `ResizeGridItemUseCase.kt` and `MoveGridItemUseCase.kt` files where these functions are called. Additionally, a condition in `resolveConflicts` was changed from `!coroutineContext.isActive` to `coroutineContext.isActive` to correctly handle conflict resolution.