Refactor: Run use cases on default dispatcher
This commit introduces the `withContext(Dispatchers.Default)` block
to ensure that the operations within the use cases are executed on the
default dispatcher. This is a best practice for CPU-intensive tasks
to avoid blocking the main thread.

Additionally, this commit wraps grid item content with a key based on
`gridItem.id` and uses `remember` for `currentGridItemSettings` and
`currentTextColor`. This change improves performance by ensuring that
these values are only recomposed when the grid item itself changes,
rather than on every recomposition.

Finally, the `pointerInput` modifier for tap gestures was removed from
`InteractiveApplicationInfoGridItem` as it was redundant with the
`detectTapGestures` within the `dragAndDropSource` block.