Refactor: Simplify application icon loading
This commit refactors the `DefaultLauncherAppsWrapper` to simplify how application icons are loaded.

Previously, the icon was fetched by calling a separate `PackageManagerWrapper`. This has been updated to directly use `applicationInfo.loadIcon(context.packageManager)`, which is a more direct and standard Android framework method.

This change removes the dependency on `PackageManagerWrapper` from `DefaultLauncherAppsWrapper`, streamlining the code and reducing an unnecessary layer of abstraction.